requirements.txt
opencv-python
pytesseract
pdf2image
Pillow
pandas
openpyxl
numpy
paddleocr
paddlepaddle
